refactor: update MyNumber class to generic class

+++ src/test/java/calculator/CalculatorSteps.java
@@ -108,12 +108,12 @@ public class CalculatorSteps {
 
 	@Then("the operation evaluates to {double}")
 	public void thenTheOperationEvaluatesTo(double val) {
-		assertEquals(val, c.eval(op), 0.0001);
+		assertEquals(val, c.eval(op).doubleValue(), 0.0001);
 	}
 
 	@Then("the operation evaluates to NaN")
 	public void thenTheOperationEvaluatesToNan() {
-		assertEquals(Double.NaN, c.eval(op), 0.0001);
+		assertEquals(Double.NaN, c.eval(op).doubleValue(), 0.0001);
 	}
 }
